\name{penAFT.plot}
\alias{penAFT.plot}
\title{Plot cross-validation curves}
\description{A function for plotting the cross-validation curves for the regularized Gehan estimator.}
\usage{
penAFT.plot(fit)
}

\arguments{
\item{fit}{A fitted model from \code{penAFT.cv}.}
}


\details{
	See \url{ajmolstad.github.io/docs/penAFT_Example.html} for usage examples.
}
