\name{penAFT.trace}
\alias{penAFT.trace}
\title{Print trace plot for the regularized Gehan estimator fit using \code{penAFT} or \code{penAFT.cv}}
\description{Print the trace plot for the regularized Gehan estimator.}
\usage{
penAFT.trace(fit, groupNames=NULL)
}

\arguments{
\item{fit}{A fitted model from \code{penAFT} orb\code{penAFT.cv}.}
\item{groupNames}{A list of groupnames to be printed when \code{fit} used penalty "SG".}
}


\details{
	See \url{ajmolstad.github.io/docs/penAFT_Example.html} for usage examples.
}
